requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Authlib==1.3.1
python-dotenv==1.0.1
PyMySQL==1.1.0
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
pytest==8.2.0
pytest-flask==1.3.0
gunicorn==21.2.0
xhtml2pdf==0.2.13
